refactor(shared-charts): move type def to typings
