+++ packages/shared/charts/src/hooks/usePlotChartDataSeries.ts
@@ -12,7 +12,7 @@ import { useEffect, useState } from "react";
 import { ensure } from "@mendix/widget-plugin-platform/utils/ensure";
 import { Datum, PlotData } from "plotly.js";
 import { executeAction } from "@mendix/widget-plugin-platform/framework/execute-action";
-import { ExtraTraceProps } from "../types";
+import { ExtraTraceProps } from "../typings/ExtraTraceProps";
 
 // Use "value" prop on EditableValue to extract AttributeValue, as AttributeValue not exported.
 type AttributeValue = EditableValue["value"];
