+++ packages/shared/charts/src/index.tsx
@@ -1,4 +1,3 @@
 export * from "./components/ChartWidget";
-export * from "./types";
 export * from "./utils/equality";
 export * from "./utils/compareAttrValuesAsc";
