+++ packages/shared/charts/src/typings/ExtraTraceProps.d.ts
